removing data and queues
