+++ README.md
@@ -5,7 +5,7 @@
 ## usage
 
 ### ```cargo make clean_run```
-This will clean and build all projects. Then run the test_game binary.
+This will clean and build all projects. Then runs the test_game binary.
 
 ### ```cargo make run```
-This will build all projects and run the test_game binary.
+This will build all projects and runs the test_game binary.
